verkefni lokið?

+++ src/lib/makeFiles.js
@@ -1,34 +1,40 @@
 import {mkdir, stat, writeFile} from 'fs/promises'
-import {join} from 'path'
 
-async function existsDir(dir) {
+export async function existsDir(dir) {
     if (!dir) {
         return false;
     }
     
     try {
         const info = await stat(dir);
-        return info.isDir(); 
+        return info.isDirectory(); 
     }
     catch(e) {
+      console.error("Error couldnt see if dir exists: ", e);
         return false;
     }
 }
 
 export async function makeDir(dir) {
     if(!await existsDir(dir)) {
-        return mkdir(dir);
+      await mkdir(dir);
     }
 }
 
 export async function makeFile(body, title, name, dir) {
-  const filePath = `./${dir}/${name}`;
+  const dirPath = `./${dir}`
+  const filePath = `${dirPath}/${name}`;
   const template = fileTemplate(title, body);
 
-  await writeFile(filePath, template, { flag: 'w+' });
+  if (existsDir(dirPath)) {
+    await writeFile(filePath, template, { flag: 'w+' });
+  }
+  else {
+    console.error("Dir does not exist: ", dir);
+  }
 }
 
-function fileTemplate(title, body) {
+export function fileTemplate(title, body) {
   return `<!doctype html>
     <html>
       <head>
@@ -36,12 +42,12 @@ function fileTemplate(title, body) {
         <meta name="viewport" content="width=device-width" />
         <link rel="stylesheet" href="styles.css" />
         <title>${title}</title>
-        <script src="script.js"></script>
       </head>
       <body>
         <main>
           ${body}
         </main>
+        <script src="script.js"></script>
       </body>
     </html>`;
 }
